src/queue.py
class Node:
    """Класс для узла очереди"""

    def __init__(self, data, next_node):
        """
        Конструктор класса Node

        :param data: данные, которые будут храниться в узле
        """
        self.data = data
        self.next_node = next_node


class Queue:
    """Класс для очереди"""

    def __init__(self):
        """Конструктор класса Queue"""
        self.head = None
        self.tail = None

    def __str__(self):
        node = self.head
        result = []
        while node is not None:
            result.append(node.data)
            node = node.next_node

        return '\n'.join(i for i in result)


    def enqueue(self, data):
        """
        Метод для добавления элемента в очередь

        :param data: данные, которые будут добавлены в очередь
        """
        if self.head is None:
            self.head = Node(data, None)
            self.tail = Node(data, None)
        else:
            node = Node(data, None)
            self.tail.next_node = node
            if self.head.next_node is None:
                self.head.next_node = node
            self.tail = node



    def dequeue(self):
        """
        Метод для удаления элемента из очереди.
        Возвращает данные удаленного элемента

        :return: данные удаленного элемента
        """
        pass


